requirements.txt
asgiref==3.8.1
bleach==6.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
crispy-bootstrap5==2024.2
Django==5.0.6
django-allauth==0.63.1
django-ckeditor==6.7.1
django-crispy-forms==2.1
django-js-asset==2.2.0
idna==3.7
Markdown==3.5.2
pillow==10.3.0
requests==2.32.2
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
webencodings==0.5.1
